server/requirements.txt
alembic==1.13.1
cryptography==41.0.7
email_validator==2.1.1
fastapi>=0.100.0
psycopg2-binary==2.9.9
pydantic>=2.0.1
pydantic-settings==2.0.3
SQLAlchemy==2.0.25
uvicorn==0.27.1
python-decouple
python-jose
passlib[bcrypt]
bcrypt==4.0.1
python-multipart
logtail-python==0.2.10
ipython